backend/requirements.txt
# Python dependencies for ConnectiLearn Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# AI and ML
openai>=1.6.1
langchain==0.0.340
langchain-openai==0.0.1

# Vector database
faiss-cpu==1.12.0
sentence-transformers==5.1.2

# PDF processing
pypdf2==3.0.1
python-docx==1.1.0
